+++ app/javascript/new_snippet.js
@@ -1,0 +1,2 @@
+import './controllers'
+import './lib/snippet_language_select'
