+++ app/javascript/show_snippet.js
@@ -1,0 +1,2 @@
+import './lib/highlight'
+import './lib/copy_to_clipboard'
