+++ app/views/layouts/application.html.erb
@@ -9,6 +9,7 @@
 
     <%= stylesheet_link_tag 'application', 'data-turbo-track': 'reload' %>
     <%= javascript_include_tag 'application', 'data-turbo-track': 'reload', defer: true %>
+    <%= yield(:custom_js) %>
   </head>
 
   <body class="h-screen bg-slate-900 text-slate-200">
