+++ webpack.config.js
@@ -5,7 +5,9 @@ module.exports = {
   mode: 'production',
   devtool: 'source-map',
   entry: {
-    application: './app/javascript/application.js'
+    application: './app/javascript/application.js',
+    new_snippet: './app/javascript/new_snippet.js',
+    show_snippet: './app/javascript/show_snippet.js',
   },
   output: {
     filename: '[name].js',
